refactor and comment example tests
