+++ Example/IlluminatorUITests/IlluminatorUITests.swift
@@ -13,47 +13,13 @@ import IlluminatorBridge
 class IlluminatorUITests: IlluminatorTestCase {
     
 
-    func test_basicWithoutIlluminator() {
-        
-        let textField = app.otherElements.containingType(.Button, identifier:"Button").childrenMatchingType(.TextField).element
-        
-        textField.tap()
-        textField.typeText("test")
-        
-        XCTAssertEqual(textField.value as? String, "test")
-        
-    }
-    
-    func test_basicWithIlluminator() {
-        
-        let interface = ExampleTestApp(testCase: self)
-        let initialState = IlluminatorTestProgress<AppTestState>.Passing(AppTestState(didSomething: false))
-        
-        let myExampleText = "test123"
-        
-        initialState
-            .apply(interface.home.enterText(myExampleText))
-            .apply(interface.home.verifyText(myExampleText))
-            .finish(self)
-    }
-
     func test_basicWithIlluminatorCompositeAction() {
-
-        let interface = ExampleTestApp(testCase: self)
-        let initialState = IlluminatorTestProgress<AppTestState>.Passing(AppTestState(didSomething: false))
-
-        let myExampleText = "test123"
-
         initialState
-            .apply(interface.home.enterAndVerifyText(myExampleText))
+            .apply(interface.home.enterAndVerifyText("test123"))
             .finish(self)
     }
     
     func test_basicWithIlluminatorAndHandlerProtocol() {
-
-        let interface = ExampleTestApp(testCase: self)
-        let initialState = IlluminatorTestProgress<AppTestState>.Passing(AppTestState(didSomething: false))
-
         let myExampleText = "test123"
 
         initialState
@@ -63,10 +29,6 @@ class IlluminatorUITests: IlluminatorTestCase {
     }
     
     func test_basicWithIlluminatorAndHandlerClosure() {
-
-        let interface = ExampleTestApp(testCase: self)
-        let initialState = IlluminatorTestProgress<AppTestState>.Passing(AppTestState(didSomething: false))
-
         let myExampleText = "test123"
 
         initialState
